Split env-server runtime out to drop lazy-import pylint disables
The render handlers need the env source, primitives, and approaches, which
created the cycle env_server -> render_policy -> agentic_approach ->
env_server and forced import-outside-toplevel lazy imports. Move the
serving loop, dispatch, render handlers, and main into a new
env_server_runtime module (the subprocess entry point) that imports the
render code at top level; env_server keeps only the import-facing API
(codecs, serialize_space, env_server_running) and imports nothing heavy, so
the main process never pulls in the cycle. env_server_running now launches
-m robocode.utils.env_server_runtime.

Also move the env_client import in mcp.server and the test imports in
test_docker_sandbox to the top level. The only remaining disable is the
intentional broad except on agent-controlled requests in the runtime
(reports errors over the wire instead of killing the connection).
